% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type I and type II error and pvalue one side
%   and two sides.R
\name{alpha.beta}
\alias{alpha.beta}
\title{Show the interaction between the probability of type 1 and type 2 error
The aim is to show the areas based on the null and alternative sampling distributions of the mean
Tests H0: mu = mu0 vs Ha: mu > mu0; H0: mu = mu0 vs Ha: mu < mu0; H0: mu = mu0 vs Ha: mu != mu0}
\usage{
alpha.beta(mu0, mua, sigmax, alpha, side, Obs = FALSE)
}
\arguments{
\item{mu0}{the parameter under the null hypothesis}

\item{mua}{the parameter under the alternative hypothesis}

\item{sigmax}{the standard deviation of the sample variance divided by the number of observations}

\item{alpha}{significance level.}

\item{side}{it can be \code{less} if it is one side test and the alternative hypothesis is less than the null hypothesis;
side it can be \code{greater} if it is one side test and the alternative hypothesis is greater than the null hypothesis;
side it can be \code{two.sided} if it is two sides test and the alternative hypothesis can be smaller than the null hypothesis or the alternative hypothesis can be greater than the null hypothesis;}

\item{Obs}{if it is \emph{TRUE} so if it is a specific value than it is possible to see on the same plot the value and the p-value}
}
\value{
Plot for the probability of type 1 and type 2 error as areas. The plots are based on Standardized Normal distributions it is also reported gamma which is the power of the test as 1-beta
}
\description{
Show the interaction between the probability of type 1 and type 2 error
The aim is to show the areas based on the null and alternative sampling distributions of the mean
Tests H0: mu = mu0 vs Ha: mu > mu0; H0: mu = mu0 vs Ha: mu < mu0; H0: mu = mu0 vs Ha: mu != mu0
}
\examples{
library(BAStat)

alpha.beta(mu0=200,mua=180,sigmax=7,alpha=0.05,Obs=183,side="left")

alpha.beta(mu0=200,mua=180,sigmax=7,alpha=0.05,side="left")
alpha.beta(mu0=200,mua=180,sigmax=7,alpha=0.1,side="left")

alpha.beta(mu0=200,mua=220,sigmax=7,alpha=0.05,Obs=218,side="right")
alpha.beta(mu0=200,mua=220,sigmax=7,alpha=0.05,side="right")
alpha.beta(mu0=200,mua=220,sigmax=7,alpha=0.1,side="right")

alpha.beta(mu0=200,mua=220,sigmax=7,alpha=0.05,side="two.sided")
alpha.beta(mu0=200,mua=180,sigmax=7,alpha=0.05,side="two.sided")
alpha.beta(mu0=200,mua=220,sigmax=7,alpha=0.05,Obs=218,side="two.sided")
alpha.beta(mu0=200,mua=180,sigmax=7,alpha=0.05,Obs=183,side="two.sided")
}
