% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type I and type II error and pvalue one side
%   and two sides binomialnormal.R
\name{alphabetabinom}
\alias{alphabetabinom}
\title{Show the interaction between the probability of type 1 and type 2 error when it is based on Binomial distribution
The aim is to show the areas based on the null and alternative sampling distributions of the proportion. Computations are without the adjusting for small samples.
Tests H0: pi = pi0 vs Ha: pi > pi0; H0: pi = pi0 vs Ha: pi < pi0; H0: pi = pi0 vs Ha: pi != pi0
It was possible to add this function to the mean function, but the purpose is also to remember that at the beginning the distribution is discrete and in the standard deviation it might be under null hypothesis or alternative hypothesis.}
\usage{
alphabetabinom(pi0, pia, n, alpha, side, Obs = FALSE)
}
\arguments{
\item{pi0}{the parameter under the null hypothesis}

\item{pia}{the parameter under the alternative hypothesis}

\item{n}{the sample size}

\item{alpha}{significance level.}

\item{side}{it can be \code{less} if it is one side test and the alternative hypothesis is less than the null hypothesis;
side it can be \code{greater} if it is one side test and the alternative hypothesis is greater than the null hypothesis;
side it can be \code{two.sided} if it is two sides test and the alternative hypothesis can be smaller than the null hypothesis or the alternative hypothesis can be greater than the null hypothesis; in the plot there is alpha divided by
two and it is represented in light blue the area and it is also specified if it is on the right side or on the left side for the comparison with the beta.}

\item{Obs}{if it is \emph{TRUE} so if it is a specific value than it is possible to see on the same plot the value and the p-value}
}
\value{
Plot for the probability of type 1 and type 2 error as areas. The plots are based on Standardized Normal distributions it is also reported gamma which is the power of the test as 1-beta
}
\description{
Show the interaction between the probability of type 1 and type 2 error when it is based on Binomial distribution
The aim is to show the areas based on the null and alternative sampling distributions of the proportion. Computations are without the adjusting for small samples.
Tests H0: pi = pi0 vs Ha: pi > pi0; H0: pi = pi0 vs Ha: pi < pi0; H0: pi = pi0 vs Ha: pi != pi0
It was possible to add this function to the mean function, but the purpose is also to remember that at the beginning the distribution is discrete and in the standard deviation it might be under null hypothesis or alternative hypothesis.
}
\examples{
library(BAStat)

alphabetabinom(pi0=0.5,pia=0.6,n=50,alpha=0.20,side="greater")
alphabetabinom(pi0=0.5,pia=0.6,n=50,alpha=0.20,Obs=33,side="greater")
alphabetabinom(pi0=0.5,pia=0.4,n=50,alpha=0.20,side="less")
alphabetabinom(pi0=0.5,pia=0.4,n=50,alpha=0.20,Obs=13,side="less")

alphabetabinom(pi0=0.5,pia=0.4,n=50,alpha=0.10,side="two.sided")
alphabetabinom(pi0=0.5,pia=0.4,n=50,alpha=0.10,Obs=16,side="two.sided")
alphabetabinom(pi0=0.5,pia=0.6,n=50,alpha=0.10,side="two.sided")
alphabetabinom(pi0=0.5,pia=0.6,n=50,alpha=0.10,Obs=33,side="two.sided")
}
