% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial distribution.R, R/probability and CDF
%   for binomial distribution.R
\name{binProbsPlots}
\alias{binProbsPlots}
\title{plot for a Binomial distribution or plot the cumulative distribution function (CDF) of a Binomial distribution.}
\usage{
binProbsPlots(
  size,
  prob,
  side,
  value.x = NULL,
  lower.x = NULL,
  upper.x = NULL,
  cdf = FALSE
)

binProbsPlots(
  size,
  prob,
  side,
  value.x = NULL,
  lower.x = NULL,
  upper.x = NULL,
  cdf = FALSE
)
}
\arguments{
\item{size}{is the sample size of the Binomial distribution;}

\item{prob}{the probability of success of the Binomial distribution;}

\item{side}{it can be \code{equal} and it the plot with the probability in one specific value that it will assigned in \code{value.x}; if side it is equal to \code{left} then the probability is considered to be on the left side or equal to the chosen value; \code{right} if the probability is on the right side or equal to the chosen value; \code{between} if it is referred to the probability between or equal to two values and in that case it has to be specified \code{lower.x} and \code{upper.x}}

\item{value.x}{the point of the probability P(X=x) or  P(X>=x) or  P(X<=x)}

\item{lower.x}{the first point of the probability in between P(xl<=X<=xu)}

\item{upper.x}{the second point of the probability in between P(xl<=X<=xu)}

\item{cdf}{if it is true then it is drawn the CDF of the Binomial distribution and on the top of the plot it is shown the difference between the extremes}
}
\value{
the distribution plot or the CDF plot and the values of the probabilities

the probability distribution or the CDF of the Binomial distribution
}
\description{
plot for a Binomial distribution or plot the cumulative distribution function (CDF) of a Binomial distribution.

Draw the the probability of the Binomial distribution and combine with the cumulative distribution function.
}
\examples{
library(BAStat)

binProbsPlots(size=12,prob=1/6,side="equal",value.x=4)#P(x=4)
binProbsPlots(size=12,prob=1/6,side="left",value.x=4)#P(x<=4)
binProbsPlots(size=12,prob=1/6,side="left",value.x=4,cdf=TRUE)#F(x=4)
binProbsPlots(size=12,prob=1/6,side="right",value.x=4)#P(x>=4)
binProbsPlots(size=12,prob=1/6,side="right",value.x=4,cdf=TRUE)#F(x>=4)
binProbsPlots(size=12,prob=1/6,side="between",lower.x=1,upper.x=4)##P(2<=x<=4)

binProbsPlots(size=12,prob=1/6,side="between",lower.x=1,upper.x=4,cdf=TRUE)#F(x=4)-F(x=1)
binProbsPlots(size=12,prob=1/6,side="between",lower.x=2,upper.x=4)##P(2<=x<=4)#in order to get the same amount as in the CDF
library(BAStat)

binProbsPlots(size=12,prob=1/6,side="equal",value.x=4)#P(x=4)
binProbsPlots(size=12,prob=1/6,side="left",value.x=4)#P(x<=4)
binProbsPlots(size=12,prob=1/6,side="left",value.x=4,cdf=TRUE)#F(x=4)
binProbsPlots(size=12,prob=1/6,side="right",value.x=4)#P(x>=4)
binProbsPlots(size=12,prob=1/6,side="right",value.x=4,cdf=TRUE)#F(x>=4)
binProbsPlots(size=12,prob=1/6,side="between",lower.x=1,upper.x=4)##P(2<=x<=4)
binProbsPlots(size=12,prob=1/6,side="between",lower.x=1,upper.x=4,cdf=TRUE)#F(x=4)-F(x=1)
binProbsPlots(size=12,prob=1/6,side="between",lower.x=2,upper.x=4)##P(2<=x<=4)#in order to get the same amount as in the CDF
}
