% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonparametric test chisq.R
\name{CAnonparam}
\alias{CAnonparam}
\title{plot decision for nonparametric test based on chi-squared with p-values and area in comparison}
\usage{
CAnonparam(Crit, Obs, df)
}
\arguments{
\item{Crit}{is the quantile on the right side of the chi-squared distribution (it can be known or computed right away)}

\item{Obs}{observed value from chi-square computation}

\item{df}{degree of freedom}
}
\value{
two plots: on the left side the chi-squared distribution and the comparison between the theoretical and observed value; on the right side it is computed the same chi-squared but focusing on the areas. In this plot on the right side there is the p-value and the theoretical alpha-value and on top there is the decision: if it can be rejected or not the null hypothesis.
}
\description{
plot decision for nonparametric test based on chi-squared with p-values and area in comparison
}
\examples{
library(BAStat)

df<-3
alpha<-0.025
Crit<-qchisq(1-alpha,df)
Obs<-5
alpha<-0.05
alpha<-0.10

CAnonparam(Crit=7.81,Obs=10,df=3)
CAnonparam(Crit=qchisq(1-alpha,df),Obs=10,df=3)
CAnonparam(Crit=qchisq(1-alpha,df),Obs=18,df=3)
CAnonparam(Crit=qchisq(1-alpha,df),Obs=8,df=3)
CAnonparam(Crit=qchisq(1-alpha,df),Obs=7,df=3)

#let's combine with a test
observations<- rbind(c(15,25,20),c(25,5,10))             
chisq.test(observations)
alpha=0.05 
CAnonparam(Crit=qchisq(1-alpha,df=2),Obs=2.5,df=2)
}
