% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdfplot.R
\name{cdfContinous}
\alias{cdfContinous}
\title{Compute the density function if you have the cumulative distribution function for a continuous variable}
\usage{
cdfContinous(Fx, lower, upper, lower.x, upper.x)
}
\arguments{
\item{Fx}{the cumulative distribution function (cdf) or the generic function to be derivated.}

\item{lower}{the numerical lower bound of the function.}

\item{upper}{the numerical upper bound of the function.}

\item{lower.x}{it can be equal to lower or a different value.}

\item{upper.x}{it can be equal to upper or a different value.}
}
\value{
On R Studio's Console: the numerical result of cdf in the bounds and the density function producted by your initial cdf
and on the plots area the plot of the region representing the computed area.
}
\description{
Compute the density function if you have the cumulative distribution function for a continuous variable
}
\examples{

library(BAStat)

cdfContinous(Fx=function(x)1/4 * x^2,lower = 0,upper=2,lower.x = 1,upper.x = 2)
cdfContinous(Fx=function(x)1-exp(-x^2),lower = 0,upper=10,lower.x = 0,upper.x = 10)#10 is a fictive limit
cdfContinous(Fx=function(x)x-1/2,lower=0,upper=2,lower.x = 1,upper.x = 1.5)
}
