% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area for normal distribution and the
%   t-student.R
\name{CdistrAreas}
\alias{CdistrAreas}
\title{Draw the area for continuous distributions: in this function is included the Standard Normal distribution and the CDF; the t-Student density, the chi-squared and F-Fisher density}
\usage{
CdistrAreas(
  distr = "Standard Normal",
  lower.x = NULL,
  upper.x = NULL,
  side,
  deg.fred,
  value.x = NULL,
  deg.fred1,
  deg.fred2,
  cdf = FALSE
)
}
\arguments{
\item{distr}{by default is intended to draw the Standardized Normal distribution, otherwise specify \emph{distr="t"} and it will be the t-distribution, but it has to be specified the degrees of freedom.
If \emph{distr="chi.sq"} it is the chi-squared between 0 and 30 be drawn. Also in this case it has to be specified the number of degrees of freedom. If \emph{distr="Fisher"} it will be density of the F-Fisher and in that case it would be needed to specify two degrees of freedom \code{deg.fred1} and \code{deg.fred2}}

\item{lower.x}{the lower bound of the symmetric and continuous distribution;}

\item{upper.x}{the upper bound of the symmetric and continuous distribution;}

\item{side}{it can be \code{left}, if the area is on the left side of the area; \code{right} if the area is on the right; \code{between} if the area is between two values}

\item{deg.fred}{is the degree of freedom and it would be required if it is the t-Student (distr="t") or chi-squared distribution (distr="chi.sq")}

\item{value.x}{it is for one side chi-squared or F-Fisher distribution}

\item{deg.fred1}{and deg.fred2 have to be specified only for the Fisher}

\item{cdf}{if it is true, then the CDF of a Standard Normal distribution will be drawn.}
}
\value{
the area under the standardized Normal distribution or the CDF for the normal or the area for the t-Student, chi-squared or F-Fisher density
}
\description{
Draw the area for continuous distributions: in this function is included the Standard Normal distribution and the CDF; the t-Student density, the chi-squared and F-Fisher density
}
\examples{
library(BAStat)

CdistrAreas(upper.x=1,side="left")#P(z<1)
CdistrAreas(upper.x=-1,side="left")#p(z<-1)

CdistrAreas(upper.x=1,side="left",cdf=T)#P(z<1)
CdistrAreas(upper.x=-1,side="left",cdf=T)#p(z<-1)

CdistrAreas(upper.x=-1,side="between")#wrong!
CdistrAreas(upper.x=-1,side="right")#wrong!

CdistrAreas(lower.x=-1,side="right")#P(z>-1)
CdistrAreas(lower.x=-1,side="right",cdf=T)#P(z>-1)
CdistrAreas(lower.x=1,side="right")#P(z>1)
CdistrAreas(lower.x=1,side="right",cdf=T)#P(z>1)

CdistrAreas(lower.x=1,upper.x=2,side="between")
CdistrAreas(lower.x=-1,upper.x=2,side="between")
CdistrAreas(lower.x=-2,upper.x=-1,side="between")

#cdf standard Normal distribution
CdistrAreas(lower.x=1,upper.x=2,side="between",cdf=T)
CdistrAreas(lower.x=-1,upper.x=2,side="between",cdf=T)
CdistrAreas(lower.x=-2,upper.x=-1,side="between",cdf=T)

CdistrAreas(lower.x=2,upper.x=-1,side="between")#wrong!
CdistrAreas(lower.x=-1,upper.x=-2,side="between")#wrong!
t-Student

CdistrAreas(distr="t",upper.x=2.49,deg.fred = 23,side="left")#P(t<2.49)
CdistrAreas(distr="t",upper.x=-2.49,deg.fred = 23,side="left")#p(t<-2.49)
CdistrAreas(distr="t",upper.x=-2.49,deg.fred = 23,side="between")#wrong!
CdistrAreas(distr="t",upper.x=-2.49,deg.fred = 23,side="right")#wrong!

CdistrAreas(distr="t",lower.x=-2.49,deg.fred = 23,side="right")#P(t>-2.49)
CdistrAreas(distr="t",lower.x=2.49,deg.fred = 23,side="right")#P(t>2.49)


CdistrAreas(distr="t",lower.x=-1.714,upper.x=1.714,deg.fred = 23,side="between") #P(-1.714<t<1.714)
CdistrAreas(distr="t",lower.x=-2.49,upper.x=3,deg.fred = 23,side="between") #P(-2.49<t<3)
CdistrAreas(distr="t",lower.x=-3,upper.x=-2.49,deg.fred = 23,side="between")#P(-3<t<-2.49)

CdistrAreas(distr="t",lower.x=2,upper.x=-1,deg.fred = 23,side="between")#wrong!
CdistrAreas(distr="t",lower.x=-1,upper.x=-2,deg.fred = 23,side="between")#wrong!

#different areas with same bounds if it is normal or t-student we get -numerically-different areas
CdistrAreas(lower.x=1,upper.x=2,side="between")
CdistrAreas(distr="t",lower.x=1,upper.x=2,deg.fred=23,side="between")

###chi.squared
CdistrAreas(distr="chi.sq",value.x =11,deg.fred=5,side="left")
CdistrAreas(distr="chi.sq",value.x=11,deg.fred=5,side="right")
CdistrAreas(distr="chi.sq",lower.x=11,upper.x=12,deg.fred=5,side="between")

###Fisher
CdistrAreas(distr="Fisher",value.x =2,deg.fred1=5,deg.fred2=10,side="left")
CdistrAreas(distr="Fisher",value.x=2.5,deg.fred1=5,deg.fred2=10,side="right")
CdistrAreas(distr="Fisher",lower.x=2,upper.x=3,deg.fred1=5,deg.fred2=10,side="between")
}
