% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe}
\alias{describe}
\title{Compute for one quantitative variable essential measurements}
\usage{
describe(x)
}
\arguments{
\item{x}{quantitative variable (continous or discrete)}
}
\description{
Compute for one quantitative variable essential measurements
}
\examples{
library(BAStat)
x<-c(25,25,26,26,27,27,28,28,29,30,30,31)
describe(x)
}
