% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe qualitative variable.R
\name{describeQL}
\alias{describeQL}
\title{The aim of this function is to show how to represent a qualitative variable as proportion or as percent: numerically and graphically.}
\usage{
describeQL(
  x,
  plot = "barplot",
  cols = c("orange", "dark green"),
  place = "",
  title_x,
  ylims = c(0, 1)
)
}
\arguments{
\item{x}{the data as a factor with labels}

\item{plot}{barplot as default diagram. It is possible to show a piechart but with the percentage.}

\item{cols}{is for the color of the bars: orange and green are on default but they can be easily changed.}

\item{place}{is for the legend to placed on top left, top right, bottom right or bottom left.}

\item{title_x}{is for the title of the plot}

\item{ylims}{is for the extreme values on the y-axis.}
}
\value{
Plot and table for the qualitative variable.
}
\description{
The aim of this function is to show how to represent a qualitative variable as proportion or as percent: numerically and graphically.
}
\examples{
library(SUdatasets)
str(smoking)
smoking$Gender<-ifelse(smoking$Gender=="1","Women","Men")
describeQL(smoking$Gender,plot="pie",title_x = "Gender",place="bottomleft")
describeQL(smoking$Gender,plot="barplot",title_x = "Gender",place="topleft")
}
