% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe quantitative variable.R
\name{describeQT}
\alias{describeQT}
\title{The aim of this function is to show how to represent a quantitative (discrete or continuous) variable numerically and graphically at the same time.}
\usage{
describeQT(
  x,
  x_lab = "",
  title_x = "",
  palette_x = "light blue",
  breaks_x = ceiling(log2(length(x)) + 1),
  plot = "histogram"
)
}
\arguments{
\item{x}{the data as a numerical vector}

\item{title_x}{is for the title of the plot}

\item{plot}{histogram as default diagram. It is possible to show a boxplot with the mean and the median on the plot.}

\item{cols}{is for the color of the bars: orange and green are on default but they can be easily changed.}

\item{place}{is for the legend to placed on top left, top right, bottom right or bottom left.}

\item{ylims}{is for the extreme values on the y-axis.}
}
\value{
Plot and table for the qualitative variable.
}
\description{
The aim of this function is to show how to represent a quantitative (discrete or continuous) variable numerically and graphically at the same time.
}
\examples{
library(SUdatasets)
str(smoking)
describeQT(smoking$Age,plot="boxplot",title_x = "Age")
describeQT(smoking$Age,plot="histogram",breaks=5,title_x = "Age")
}
