% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exy.R
\name{Exy}
\alias{Exy}
\title{Expected value of product of random variables E(xy)}
\usage{
Exy(xx, yy, pxxyy)
}
\arguments{
\item{xx}{the first numerical variable, x}

\item{yy}{the second numerical variable, y}

\item{pxxyy}{the distribution function between the two variables; it has to be presented as a matrix, even rectangular}
}
\value{
the expected value of product of random variables: E(XY)=sum_i sum_j x_i y_j p(x_iy_j)
}
\description{
This function is useful for the covariance and correlation between two random variables. It is not substituting the command \code{cov} or \code{cor} but integrating them.
}
\examples{
library(BAStat)
pxy<-matrix(
c(0.1, 0.05,  0.00, 0.15,
 0.05, 0.10,  0.15, 0.10,
 0.10, 0.05,  0.10, 0.05),
nrow=4,ncol=3)
y<-c(3,5,4)
x<-c(2,4,6,8)
Exy(xx=x,yy=y,pxxyy=pxy)
}
