% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{histog}
\alias{histog}
\title{Histogram for one variable}
\usage{
histog(x, x_lab = "", title_x = "", palette_x = "light blue", box = FALSE)
}
\arguments{
\item{x}{quantitative variable (continous or discrete)}

\item{x_lab}{specify the name of the variable on x-axis}

\item{palette_x}{define the color of your histogram or by default is "light blue"}

\item{box}{if it is true it can produce a boxplot as well
As a result the histogram is with the Normal density curve and the kernel density from the observations}
}
\description{
Histogram for one variable
}
\examples{
library(BAStat)
x<-c(25,25,26,26,27,27,28,28,29,30,30,31)
histog(x)
}
