% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indipendence btw two stok var.R
\name{indip}
\alias{indip}
\title{Indipendence between two random variables}
\usage{
indip(xx, yy, pxxyy)
}
\arguments{
\item{xx}{the first numerical variable, x}

\item{yy}{the second numerical variable, y}

\item{pxxyy}{the distribution function between the two variables; it has to be presented as a matrix, even rectangular}
}
\value{
the thoretical distribution function obtained by multiplying the marginal distributions: p(x=x_iy=y_j)=p(x=x_i)*p(y=y_j)
}
\description{
Indipendence between two random variables
}
\examples{
library(BAStat)
y<-c(6,7,8)
x<-c(10,20)
pxy<-matrix(
 c(0.1,0.2,0.1,0.2,0.3,0.1),
 nrow=2,ncol=3) 
}
