% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integral and plot the area.R
\name{integPlot}
\alias{integPlot}
\title{Compute defined integral and draw the area that is referred to}
\usage{
integPlot(fx, lower, upper)
}
\arguments{
\item{fx}{the integrand, the mathematical function to be integrated.}

\item{lower}{the numerical lower bound of the integral.}

\item{upper}{the numerical upper bound of the integral.}
}
\value{
On R Studio's Console: the numerical result of the integral and on the plots area the plot of the region representing the computed area.
}
\description{
The goal of this function is to compute defined integral and to show the corresponding area.
Note that this function can be adapted also for indefinite integrals (integrals between -Inf and Inf) but a tolerance equal to 1e-06 is introduced.
}
\examples{

library(BAStat)

integrand<-function(x){3*x^4}
integPlot(fx=integrand,lower = 0,upper = 2)
integrand<-function(x){x*exp(x)}
integPlot(fx=integrand,lower = 0,upper = 1)

integrand <- function(x) {1/sqrt(2*pi)*exp(-x^2/2)} #this is the Normal distribution
integPlot(fx=integrand,lower = -1.96,upper = 1.96)

integPlot(fx=function(x)1/4 * x^2,lower = 0,upper = 2)
integPlot(fx=function(x)1/4 * x^2,lower = -Inf,upper = Inf)# indefinite integral
integPlot(fx=function(x)1/4 * x^2,lower = -Inf,upper = 2) # improper integral: one of the bounds is infinity
integPlot(fx=function(x)1/4 * x^2,lower = 0,upper = Inf)# improper integral 
}
