% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/the observed value for test with 1 parameter.R
\name{observed}
\alias{observed}
\title{This small function is for standardizing the observed value before the test for one parameter}
\usage{
observed(x = NULL, parameter = NULL, binom = F, n = NULL)
}
\arguments{
\item{x}{the numerical variable to be standardized}

\item{parameter}{the mean or the proportion}

\item{binom}{if it is TRUE than the observed value will be the proportion from the observations}

\item{n}{number of observations; it would be needed only if it is supposed to refer to the Binomial distribution}
}
\value{
the standardized observed value for the test of one parameter
}
\description{
There are several functions in R for those computations and in particular for the proportion there are many formulas for the correction to normality and decentralizing the attention to the the initial formula and the power of the central limit theorem.
}
\examples{
library(BAStat)

ex<-c(2970,3020,3005,2900,2940,2925)
observed(ex,parameter=3000)

x<-10
n<-160
parameter<-0.08
observed(x=10,parameter=0.08,binom=T,n=160)
}
