% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{t.pvalues}
\alias{t.pvalues}
\title{P-value for t distribution}
\usage{
t.pvalues(tCrit, tObs,deg.freed,side)
}
\arguments{
\item{tCrit}{Critical value}
\item{tObs}{Observed value}
\item{deg.freed}{degrees of freedom}
\item{side}{if the test is on the right-side, "R";
if the test is on the left-side, "L"; 
if the test is on both sides, "B"}
}


\examples{
library(BAStat)
t.pvalues(tCrit = qt(0.975,28),tObs= 3.09,deg.freed=28,side="R")
}
