% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvaluetsides.R
\name{t.pvalues}
\alias{t.pvalues}
\title{Compute p-values and take decisions under a t-distribution}
\usage{
\method{t}{pvalues}(tCrit, tObs, degree.freed, alternative)
}
\arguments{
\item{tCrit}{critical value}

\item{tObs}{observed value already studentised}

\item{degree.freed}{it is the number of degree of freedom}

\item{side}{\code{B} if it is the alternative hypothesis on both sides.}
}
\value{
plot with the critical region under the t-Student hypothesis and the p-value
}
\description{
Compute p-values and take decisions under a t-distribution
}
\examples{
library(BAStat)
tCrit =qt(0.95,18)
tCrit =qt(0.975,18)
tCrit =qt(0.025,18)
pt(3.0866537,18)
qt(0.003181154,18)
t.pvalues(tCrit = 1.734064,tObs= 3.0866537,degree.freed=18,alternative="greater")
t.pvalues(tCrit = -1.734064,tObs= -3.0866537,degree.freed=18,alternative="less")
t.pvalues(tCrit = -2.100922,tObs= -3.0866537,degree.freed=18,alternative="two.sided")
t.pvalues(tCrit = 2.100922,tObs= 3.0866537,degree.freed=18,alternative="two.sided")
}
