% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvaluetsides.R
\name{tPvalues}
\alias{tPvalues}
\title{Compute p-values and take decisions under a t-distribution}
\usage{
tPvalues(tCrit, tObs, degree.freed, alternative)
}
\arguments{
\item{tCrit}{critical value}

\item{tObs}{observed value already studentised}

\item{degree.freed}{it is the number of degree of freedom}

\item{alternative}{\code{two.sided} if the alternative hypothesis is on both sides; \code{greater} if the alternative hypothesis is on the right side; \code{less} if the alternative hypothesis is on the left side}
}
\value{
plot with the comparison between the theoretical area for the critical region and the obtained p-value. On the top of the plot is reported the answer: if it has to be rejected of not the null hypothesis. On the left side of the plot is reported the computed p-value and the computed area for the theoretical t-value.
}
\description{
This function helps in the learning process: when the null hypothesis can be rejected based on the p-value. It is implemented for one and two tailed t-tests. It is for the test of hypothesis for the mean of one or two populations as well as a support for the estimated parameters of a regression model.
}
\examples{
library(BAStat)
tCrit =qt(0.95,18)
tPvalues(tCrit = 1.734064,tObs= 3.0866537,degree.freed=18,alternative="greater")
tCrit =qt(0.05,18)
tPvalues(tCrit = -1.734064,tObs= -3.0866537,degree.freed=18,alternative="less")
tCrit =qt(0.025,18)
tPvalues(tCrit = -2.100922,tObs= -3.0866537,degree.freed=18,alternative="two.sided")
tCrit =qt(0.975,18)
tPvalues(tCrit = 2.100922,tObs= 3.0866537,degree.freed=18,alternative="two.sided")
}
