% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type I and type II error and pvalue one side
%   and two sides.R
\name{alpha.beta}
\alias{alpha.beta}
\title{Show the interaction between the probability of type 1 and type 2 error
Tests H0: mu = mu0 vs Ha: mu > mu0; H0: mu = mu0 vs Ha: mu < mu0; H0: mu = mu0 vs Ha: mu != mu0}
\usage{
alpha.beta(mu0, mua, sigmax, alpha, side, Obs = FALSE)
}
\arguments{
\item{mu0}{the parameter under the null hypothesis}

\item{mua}{the parameter under the alternative hypothesis}

\item{sigmax}{the standard deviation of the sample variance divided by the number of observations}

\item{alpha}{significance level.}

\item{side}{it can be \code{right} if it is one side test and the alternative hypothesis is greater than the null hypothesis;}

\item{step}{if it is \emph{"4: t-test for two pop. means"} it is the fourth step and it is based on t-test; the t-test is provided with equal variances (pooled variance in t-statistics) if it was not rejected the null hypothesis at the third step; the t-test for unequal variances (Welch test) if it was rejected the null hypothesis at the third step.}
}
\value{
four steps: 1-descriptive statistics for each group; 2-confidence interval for the mean and the variance for each group; 3-test F check if the variances can be assumed equal or not; 4-t-test based on the assumption or not of equal variances. For step 3 and 4 are drawn plots in order to reject the null hypothesis
}
\description{
Show the interaction between the probability of type 1 and type 2 error
Tests H0: mu = mu0 vs Ha: mu > mu0; H0: mu = mu0 vs Ha: mu < mu0; H0: mu = mu0 vs Ha: mu != mu0
}
\examples{
library(BAStat)

alpha.beta(mu0=200,mua=180,sigmax=7,alpha=0.05,Obs=183,side="left")

alpha.beta(mu0=200,mua=180,sigmax=7,alpha=0.05,side="left")
alpha.beta(mu0=200,mua=180,sigmax=7,alpha=0.1,side="left")

alpha.beta(mu0=200,mua=220,sigmax=7,alpha=0.05,Obs=218,side="right")
alpha.beta(mu0=200,mua=220,sigmax=7,alpha=0.05,side="right")
alpha.beta(mu0=200,mua=220,sigmax=7,alpha=0.1,side="right")

alpha.beta(mu0=200,mua=220,sigmax=7,alpha=0.05,side="two.sided")
alpha.beta(mu0=200,mua=180,sigmax=7,alpha=0.05,side="two.sided")
alpha.beta(mu0=200,mua=220,sigmax=7,alpha=0.05,Obs=218,side="two.sided")
alpha.beta(mu0=200,mua=180,sigmax=7,alpha=0.05,Obs=183,side="two.sided")
}
