% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area for normal distribution.R
\name{area}
\alias{area}
\title{Draw the area under a standardized Normal density curve}
\usage{
area(
  lower.x = -4,
  upper.x = 4,
  mu = 0,
  sigma = 1,
  col = "light blue",
  density = NULL,
  titles = "Area"
)
}
\arguments{
\item{lower.x}{the lower bound for a standardized Normal distribution, by default it is equal to -4}

\item{upper.x}{the upper bound for a standardized Normal distribution, by default it is equal to 4}

\item{mu}{the population mean, by default is equal to zero}

\item{sigma}{the standarddeviation of the Normal; by default it is equal to 1}

\item{col}{it is the color of the area under the curve,  by default it is light blue: it can be changed}

\item{density}{draw the kernel density function from the data}

\item{titles}{it the title on the top of the plot}
}
\value{
the area under the standardized Normal distribution
}
\description{
Draw the area under a standardized Normal density curve
}
\examples{
library(BAStat)
#P(-1<Z<1)
area(lower.x=-1, upper.x=1, mu=0, sigma=1,col="light blue",titles="Area")
#P(Z<1)
area(upper.x=1)
#P(Z>1)
area(lower.x = 1)
}
