% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area for normal distribution.R
\name{area}
\alias{area}
\title{Draw the area under the Normal density curve}
\usage{
area(
  lower.x = -4,
  upper.x = 4,
  mu = 0,
  sigma = 1,
  col = "light blue",
  density = NULL,
  titles = "Area"
)
}
\arguments{
\item{lower.x}{the lower bound for a Normal distribution, by default it is equal to -4}

\item{upper.x}{the upper bound for a Normal distribution, by default it is equal to 4}

\item{mu}{the population mean, by default is equal to zero}

\item{sigma}{the standarddeviation of the Normal; by default it is equal to 1}

\item{col}{it is the color of the area under the curve,  by default it is light blue: it can be changed}

\item{titles}{it the title on the top of the plot}
}
\value{
the area under the standardized Normal distibution
}
\description{
Draw the area under the Normal density curve
}
\examples{
library(BAStat)
#P(-4<Z<4)
area(lower.x=-4, upper.x=4, mu=0, sigma=1,col="light blue",titles="Area")
#P(Z<2)
area(upper.x=2)
#P(Z>3)
area(lower.x=3)
#P(X>1) and X is a Normal distribution with mean equal to 3 and variance equal to 2
area(lower.x=1,mu=3,sigma=sqrt(2))
}
