% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial distribution.R, R/probability and CDF
%   for binomial distribution.R
\name{binProbsPlots}
\alias{binProbsPlots}
\title{plot a Binomial distribution or the CDF for a Binomial distribution.}
\usage{
binProbsPlots(
  size,
  prob,
  side,
  value.x = NULL,
  lower.x = NULL,
  upper.x = NULL,
  cdf = FALSE
)

binProbsPlots(
  size,
  prob,
  side,
  value.x = NULL,
  lower.x = NULL,
  upper.x = NULL,
  cdf = FALSE
)
}
\arguments{
\item{size}{is the sample size of the Binomial distribution;}

\item{prob}{the probability of success of the Binomial distribution;}

\item{side}{it can be \code{equal} and it the plot with the probability in one specific value that it will assigned in \code{value.x}; if side it is equal to \code{left} then the probability is considered to be on the left side or equal to the chosen value; \code{right} if the probability is on the right side or equal to the chosen value; \code{between} if it is referred to the probability between or equal to two values and in that case it has to be specified \code{lower.x} and \code{upper.x}}

\item{cdf}{if it is true then it is drawn the CDF of the Binomial distribution and on the top of the plot it is shown the difference between the extremes}

\item{distr}{by default is intended to use the Normal distribution, otherwise specify \emph{distr="t"} for the t-distribution or \emph{distr="binom"} for the Binomial distribution.}

\item{alpha}{the confidence level;}

\item{center}{the mean of the sample for the Normal or t-distribution, the number of successes divided by the number of observations for the proportion}

\item{stddev}{the standard deviation of the sample for the t-distribution; or the standard deviation of the population in case of normality or the use of central limit theorem for the normal approximation; the standard deviation based on the estimated proportion}

\item{n}{number of observations}

\item{degree.fred}{for the t-distribution: specify the number of degrees of freedom}
}
\value{
the confidence level, the statistical marginal error, the lower and upper bound of the confidence interval. In case of proportion and the number of observations is greater than 50, it is also added the Yates' continuity. In the plots area are also automatically simulated and shown the 200 sampled confidence interval in order to remind that the parameter of the population is unknown and it is only possible to guess with a certain confidence where the parameter should be located

the probability distribution or the CDF of the Binomial distribution
}
\description{
plot a Binomial distribution or the CDF for a Binomial distribution.

Draw the the probability of the Binomial distribution and combine with the cumulative distribution function.
}
\examples{
library(BAStat)

binProbsPlots(size=12,prob=1/6,side="equal",value.x=4)#P(x=4)
binProbsPlots(size=12,prob=1/6,side="left",value.x=4)#P(x<=4)
binProbsPlots(size=12,prob=1/6,side="left",value.x=4,cdf=TRUE)#F(x=4)
binProbsPlots(size=12,prob=1/6,side="right",value.x=4)#P(x>=4)
binProbsPlots(size=12,prob=1/6,side="right",value.x=4,cdf=TRUE)#F(x>=4)
binProbsPlots(size=12,prob=1/6,side="between",lower.x=1,upper.x=4)##P(2<=x<=4)

binProbsPlots(size=12,prob=1/6,side="between",lower.x=1,upper.x=4,cdf=TRUE)#F(x=4)-F(x=1)
binProbsPlots(size=12,prob=1/6,side="between",lower.x=2,upper.x=4)##P(2<=x<=4)#in order to get the same amount as in the CDF
library(BAStat)

binProbsPlots(size=12,prob=1/6,side="equal",value.x=4)#P(x=4)
binProbsPlots(size=12,prob=1/6,side="left",value.x=4)#P(x<=4)
binProbsPlots(size=12,prob=1/6,side="left",value.x=4,cdf=TRUE)#F(x=4)
binProbsPlots(size=12,prob=1/6,side="right",value.x=4)#P(x>=4)
binProbsPlots(size=12,prob=1/6,side="right",value.x=4,cdf=TRUE)#F(x>=4)
binProbsPlots(size=12,prob=1/6,side="between",lower.x=1,upper.x=4)##P(2<=x<=4)
binProbsPlots(size=12,prob=1/6,side="between",lower.x=1,upper.x=4,cdf=TRUE)#F(x=4)-F(x=1)
binProbsPlots(size=12,prob=1/6,side="between",lower.x=2,upper.x=4)##P(2<=x<=4)#in order to get the same amount as in the CDF
}
