% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence interval for one population.R
\name{CI}
\alias{CI}
\title{Confidence intervals for the mean of one population supposed to be Normal, with known and unknown variance and Binomial. There are several functions in R for those computations and in particular for the proportion there are several formulas for correction to the normality.}
\usage{
CI(distr = "Standard Normal", alpha, center, stddev, n, degree.fred)
}
\arguments{
\item{distr}{by default is intended to use the Normal distribution, otherwise specify \emph{distr="t"} for the t-distribution or \emph{distr="binom"} for the Binomial distribution.}

\item{alpha}{the confidence level;}

\item{center}{the mean of the sample for the Normal or t-distribution, the number of successes divided by the number of observations for the proportion}

\item{stddev}{the standard deviation of the sample for the t-distribution; or the standard deviation of the population in case of normality or the use of central limit theorem for the normal approximation; the standard deviation based on the estimated proportion}

\item{n}{number of observations}

\item{degree.fred}{for the t-distribution: specify the number of degrees of freedom}
}
\value{
the confidence level, the statistical marginal error, the lower and upper bound of the confidence interval. In case of proportion and the number of observations is greater than 50, it is also added the Yates' continuity. In the plots area are also automatically simulated and shown the 200 sampled confidence interval in order to remind that the parameter of the population is unknown and it is only possible to guess with a certain confidence where the parameter should be located
}
\description{
Confidence intervals for the mean of one population supposed to be Normal, with known and unknown variance and Binomial. There are several functions in R for those computations and in particular for the proportion there are several formulas for correction to the normality.
}
\examples{
library(BAStat)

CI(alpha=0.05,center=12,stddev=3,n=30)#in case of Normal distribution
CI(distr="t",alpha=0.05,center=12,stddev=3,n=15,degree.fred=15-1) #for the t-distribution
phat<-7/12#number of success divided by the number of observations
CI(distr="binom",alpha=0.05,center=phat,stddev=0.1423,n=12) #for the binomial and n is less than 50
x<-70 #number of success
n<-120 #number of observations

stddev<-sqrt((x/n*(1-x/n)/n))
CI(distr="binom",alpha=0.05,center=70/120,stddev=stddev,n=120)
}
