% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvaluetsides.R
\name{t.pvalues}
\alias{t.pvalues}
\title{Compute p-values and take decisions under a t-distribution}
\usage{
\method{t}{pvalues}(tCrit, tObs, deg.freed, side)
}
\arguments{
\item{tCrit}{critical value}

\item{tObs}{observed value already studentised}

\item{deg.freed}{it is the number of degree of freedom}

\item{side}{\code{B} if it is the alternative hypothesis on both sides.}
}
\value{
plot with the critical region under the t-Student hypothesis and the p-value
}
\description{
Compute p-values and take decisions under a t-distribution
}
\examples{
library(BAStat)
t.pvalues(tCrit=qt(0.95,6), tObs=2,deg.freed=6,side="B")
}
